# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/resource/cronjob/list/{namespace}": {
            "get": {
                "description": "获取某一用户空间下的所有 CronJob 对象",
                "tags": [
                    "resource"
                ],
                "summary": "获取某一用户空间下的所有 CronJob 对象",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户的命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
        "/resource/job/detail/{name}/{namespace}": {
            "get": {
                "description": "查询某一Job对象的详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "查询某一Job对象的详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job对象名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户的命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200, \"message\":\"OK\", \"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
        "/resource/persistentvolumeclaim/list/{namespace}": {
            "get": {
                "description": "获取某一用户创建的所有PersistentVolumeClaim对象",
                "tags": [
                    "resource"
                ],
                "summary": "获取某一用户创建的所有PersistentVolumeClaim对象",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户的命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/resource/pod/container/{podId}/{namespace}": {
            "get": {
                "description": "获取某一 Pod 中的所有容器对象.",
                "tags": [
                    "resource"
                ],
                "summary": "获取某一 Pod 中的所有容器对象.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pod ID",
                        "name": "podId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/resource/pod/detail/{name}/{namespace}": {
            "get": {
                "description": "查询某一 Pod 对象的详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "查询某一 Pod 对象的详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pod 对象名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200, \"message\":\"OK\", \"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
        "/resource/pod/list/{namespace}": {
            "get": {
                "description": "获取某一命名空间下的所有 Pod 对象",
                "tags": [
                    "resource"
                ],
                "summary": "获取某一命名空间下的所有 Pod 对象",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"message\":\"OK\",\"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
        "/resource/service/detail/{name}/{namespace}": {
            "get": {
                "description": "查询某一 Service 对象的详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "resource"
                ],
                "summary": "查询某一 Service 对象的详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service 对象名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户的命名空间",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200, \"message\":\"OK\", \"data\":{\"\"}}",
                        "schema": {
                            "$ref": "#/definitions/tool.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "tool.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}